Refactor storeMarker
